+++ rapports.py
@@ -11,7 +11,7 @@ Auteur        : Pierre Théberge
 Compagnie     : Innovations, Performances, Technologies inc.
 Créé le       : 2025-08-05
 Modifié le    : 2026-09-24
-Version       : 0.6.2
+Version       : 0.6.3
 Copyright     : Pierre Théberge
 
 Description
@@ -144,6 +144,7 @@ Modifications
                                 erreurs JS du navigateur en DEBUG.
 0.6.1  - 2026-09-24   ES-28   : Synchronisation de version (aucun changement fonctionnel).
 0.6.2  - 2026-09-24   ES-28   : Synchronisation de version (aucun changement fonctionnel).
+0.6.3  - 2026-09-24   ES-28   : Synchronisation de version (aucun changement fonctionnel).
 
 Paramètres
 ----------